Initial work for new Serial display type
This is for cases where you have many pages with same template, but don't
want to create an actual, separate Page for each. Possible example is news posts.
Each post on a Serial display will be uniquely accessible, using the Page's
"link" as a prefix (ex: "posts/"), and some uid (ex: "12-feb-2012_0530.html") as
the page name.